just opening random files, and fixing, fixing crap .. can't say what exactly today. Just some working commit.

+++ src/glcam.h
@@ -64,7 +64,7 @@ private:
 
     Item_cam* cam;
     QPoint startpoint;
-    float zoom;
+    GLdouble zoom;
     GLuint object;
 
     GLfloat readpixel[4];
